[LPT] Transpose transformation & Reshape transformation

+++ inference-engine/tests/functional/plugin/shared/src/low_precision_transformations/relu_transformation.cpp
@@ -42,8 +42,8 @@ InferenceEngine::Blob::Ptr ReluTransformation::GenerateInput(const InferenceEngi
 
     return FuncTestUtils::createAndFillBlobConsistently(
         info.getTensorDesc(),
-        fqOnData.empty() ? 25.f : fqOnData.outputHighValues[0] - fqOnData.outputLowValues[0],
-        fqOnData.empty() ? -12.5f : fqOnData.outputLowValues[0],
+        static_cast<uint32_t>(fqOnData.empty() ? 25.f : fqOnData.outputHighValues[0] - fqOnData.outputLowValues[0]),
+        static_cast<int32_t>(fqOnData.empty() ? -12.5f : fqOnData.outputLowValues[0]),
         1ul);
 }
